refactor(orchestrator): extract AtomicPromptBuilder + LeafDispatcher from _execute_atomic_ac
R4 (Phase 5 structural refactor). Pure structural move — zero behaviour change,
tests unmodified.

- AtomicPromptBuilder (atomic_prompt_builder.py): the prompt/task-section
  assembly — label/indent, governed task section, success-contract block,
  retry/parallel-awareness sections, cwd scan, completion contract. Also hosts
  _build_success_contract_block (re-exported from parallel_executor for tests).
- LeafDispatcher.stream() (leaf_dispatcher.py): the stall-scoped runtime
  dispatch + per-message streaming/heartbeat loop. Mutates a shared
  LeafDispatchState so the executor's except/finally observe the same mid-loop
  runtime handle / session id / partial messages they did inline — including on
  the exception path.

_execute_atomic_ac keeps the same signature and stays the orchestrator
(runtime-handle lifecycle, effort/memory prep, verifier/evidence post-processing,
teardown); it shrinks 716 → ~410 lines.
