+++ src/ouroboros/orchestrator/leaf_dispatcher.py
@@ -1,0 +1,268 @@
+"""Runtime dispatch + streaming/heartbeat consumption for an atomic leaf.
+
+Extracted verbatim from ``ParallelACExecutor._execute_atomic_ac`` (work order
+R4). This module owns the stall-scoped runtime dispatch and the per-message
+streaming loop: the resettable stall ``CancelScope``, runtime-handle threading,
+recovery/lifecycle event emission, heartbeat emission, projected-message
+persistence, and tool/thinking event emission.
+
+Stall/heartbeat timing is subtle, so the extraction is a pure structural move:
+every await point, deadline reset, exception path, and event emission stays in
+exactly the same relative order it had inline. The mutable loop state
+(``messages``, ``runtime_handle``, ``ac_session_id``, ...) lives on the shared
+:class:`LeafDispatchState` the executor passes in, so the executor's ``except``
+and ``finally`` observe the same mid-loop values they did when the loop body was
+inline — including on the exception path, where the latest runtime handle and
+partial message list must remain visible for teardown.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, replace
+import time
+from typing import TYPE_CHECKING, Any
+
+import anyio
+
+from ouroboros.orchestrator.adapter import AgentMessage, RuntimeHandle
+from ouroboros.orchestrator.evidence.runtime_metadata import (
+    HEARTBEAT_INTERVAL_SECONDS,
+    STALL_TIMEOUT_SECONDS,
+)
+from ouroboros.orchestrator.runtime_message_projection import project_runtime_message
+
+if TYPE_CHECKING:
+    from ouroboros.orchestrator.execution_runtime_scope import (
+        ACRuntimeIdentity,
+        ExecutionNodeIdentity,
+    )
+    from ouroboros.orchestrator.parallel_executor import ParallelACExecutor
+
+
+@dataclass
+class LeafDispatchState:
+    """Mutable streaming state shared between the executor and the dispatcher.
+
+    The executor seeds this with the pre-dispatch runtime handle and its own
+    ``messages`` list (by reference), then reads the mutated fields after the
+    stream — and, critically, from within its ``except``/``finally`` when the
+    runtime raises mid-stream.
+    """
+
+    messages: list[AgentMessage]
+    runtime_handle: RuntimeHandle | None
+    ac_session_id: str | None = None
+    message_count: int = 0
+    final_message: str = ""
+    success: bool = False
+    stalled: bool = False
+
+
+class LeafDispatcher:
+    """Dispatch one atomic leaf to the runtime and consume its message stream."""
+
+    def __init__(self, executor: ParallelACExecutor) -> None:
+        self._executor = executor
+
+    async def stream(
+        self,
+        *,
+        state: LeafDispatchState,
+        prompt: str,
+        tools: list[str],
+        system_prompt: str,
+        execute_effort_kwargs: dict[str, Any],
+        runtime_identity: ACRuntimeIdentity,
+        execution_context_id: str,
+        session_id: str,
+        ac_index: int,
+        ac_content: str,
+        is_sub_ac: bool,
+        parent_ac_index: int | None,
+        sub_ac_index: int | None,
+        node_identity: ExecutionNodeIdentity | None,
+        retry_attempt: int,
+        label: str,
+        indent: str,
+        execution_counters: dict[str, int] | None,
+    ) -> None:
+        """Run the stall-scoped dispatch loop, mutating ``state`` in place."""
+        executor = self._executor
+
+        lifecycle_event_type = (
+            "execution.session.resumed"
+            if executor._is_resumable_runtime_handle(state.runtime_handle)
+            else "execution.session.started"
+        )
+        lifecycle_emitted = False
+        emitted_recovery_turn_ids: set[str] = set()
+
+        # Stall detection: CancelScope with resettable deadline (RC6)
+        last_heartbeat = time.monotonic()
+        exec_start = time.monotonic()
+
+        with anyio.CancelScope(
+            deadline=anyio.current_time() + STALL_TIMEOUT_SECONDS,
+        ) as stall_scope:
+            async for message in executor._adapter.execute_task(
+                prompt=prompt,
+                tools=tools,
+                system_prompt=system_prompt,
+                resume_handle=state.runtime_handle,
+                **execute_effort_kwargs,
+            ):
+                # Reset stall deadline on every message (RC6 core)
+                stall_scope.deadline = anyio.current_time() + STALL_TIMEOUT_SECONDS
+                if message.resume_handle is not None:
+                    state.runtime_handle = executor._remember_ac_runtime_handle(
+                        ac_index,
+                        message.resume_handle,
+                        execution_context_id=execution_context_id,
+                        is_sub_ac=is_sub_ac,
+                        parent_ac_index=parent_ac_index,
+                        sub_ac_index=sub_ac_index,
+                        node_identity=node_identity,
+                        retry_attempt=retry_attempt,
+                    )
+
+                if state.runtime_handle is not None and state.runtime_handle.native_session_id:
+                    state.ac_session_id = state.runtime_handle.native_session_id
+                elif (
+                    message.resume_handle is None
+                    and isinstance(message.data.get("session_id"), str)
+                    and message.data["session_id"]
+                ):
+                    state.ac_session_id = message.data["session_id"]
+
+                state.runtime_handle = executor._with_native_session_id(
+                    state.runtime_handle, state.ac_session_id
+                )
+                if state.runtime_handle is not None and message.resume_handle is not None:
+                    message = replace(message, resume_handle=state.runtime_handle)
+
+                recovery_discontinuity = executor._runtime_recovery_discontinuity(
+                    state.runtime_handle
+                )
+                if recovery_discontinuity is not None:
+                    replacement = recovery_discontinuity.get("replacement", {})
+                    replacement_turn_id = replacement.get("turn_id")
+                    if isinstance(replacement_turn_id, str) and replacement_turn_id:
+                        if replacement_turn_id not in emitted_recovery_turn_ids:
+                            await executor._emit_ac_runtime_event(
+                                event_type="execution.session.recovered",
+                                runtime_identity=runtime_identity,
+                                ac_content=ac_content,
+                                runtime_handle=state.runtime_handle,
+                                execution_id=execution_context_id,
+                                session_id=state.ac_session_id,
+                            )
+                            emitted_recovery_turn_ids.add(replacement_turn_id)
+
+                state.messages.append(message)
+                state.message_count += 1
+                if execution_counters is not None:
+                    async with executor._execution_counters_lock:
+                        execution_counters["messages_count"] = (
+                            execution_counters.get("messages_count", 0) + 1
+                        )
+
+                # RC1: Emit heartbeat piggybacking on message flow
+                now = time.monotonic()
+                if now - last_heartbeat >= HEARTBEAT_INTERVAL_SECONDS:
+                    await executor._event_emitter.emit_heartbeat(
+                        session_id=session_id,
+                        ac_index=ac_index,
+                        ac_id=runtime_identity.ac_id,
+                        elapsed_seconds=now - exec_start,
+                        message_count=state.message_count,
+                        node_identity=node_identity,
+                    )
+                    last_heartbeat = now
+
+                projected = project_runtime_message(message)
+
+                persisted_session_id = executor._runtime_resume_session_id(state.runtime_handle)
+                if not lifecycle_emitted and persisted_session_id:
+                    await executor._emit_ac_runtime_event(
+                        event_type=lifecycle_event_type,
+                        runtime_identity=runtime_identity,
+                        ac_content=ac_content,
+                        runtime_handle=state.runtime_handle,
+                        execution_id=execution_context_id,
+                        session_id=persisted_session_id,
+                    )
+                    lifecycle_emitted = True
+                    executor._remember_ac_runtime_handle(
+                        ac_index,
+                        state.runtime_handle,
+                        execution_context_id=execution_context_id,
+                        is_sub_ac=is_sub_ac,
+                        parent_ac_index=parent_ac_index,
+                        sub_ac_index=sub_ac_index,
+                        node_identity=node_identity,
+                        retry_attempt=retry_attempt,
+                    )
+
+                session_tool_event = executor._build_session_tool_called_event(
+                    session_id,
+                    projected=projected,
+                )
+                if session_tool_event is not None:
+                    await executor._event_store.append(session_tool_event)
+
+                if executor._should_emit_session_progress_event(
+                    message,
+                    projected=projected,
+                    messages_processed=len(state.messages),
+                ):
+                    session_progress_event = executor._build_session_progress_event(
+                        session_id,
+                        message,
+                        projected=projected,
+                    )
+                    await executor._event_store.append(session_progress_event)
+
+                if projected.is_tool_call and projected.tool_name is not None:
+                    # RC6: Tool invocations prove liveness — reset stall
+                    # deadline so long-running tools (Bash, external APIs)
+                    # are not falsely detected as stalls.
+                    stall_scope.deadline = anyio.current_time() + STALL_TIMEOUT_SECONDS
+                    if execution_counters is not None:
+                        async with executor._execution_counters_lock:
+                            execution_counters["tool_calls_count"] = (
+                                execution_counters.get("tool_calls_count", 0) + 1
+                            )
+                    tool_input = projected.tool_input
+                    tool_detail = executor._format_tool_detail(projected.tool_name, tool_input)
+                    executor._console.print(f"{indent}[yellow]{label} → {tool_detail}[/yellow]")
+                    executor._flush_console()
+
+                    await executor._event_emitter.emit_atomic_tool_started(
+                        runtime_identity=runtime_identity,
+                        tool_name=projected.tool_name,
+                        tool_detail=tool_detail,
+                        tool_input=tool_input,
+                        runtime_metadata=executor._runtime_event_metadata(message),
+                    )
+
+                if projected.is_tool_result and projected.tool_name is not None:
+                    await executor._event_emitter.emit_atomic_tool_completed(
+                        runtime_identity=runtime_identity,
+                        tool_name=projected.tool_name,
+                        tool_result_text=projected.content,
+                        runtime_metadata=executor._runtime_event_metadata(message),
+                    )
+
+                if projected.thinking:
+                    await executor._event_emitter.emit_atomic_thinking(
+                        runtime_identity=runtime_identity,
+                        thinking_text=projected.thinking,
+                        runtime_metadata=executor._runtime_event_metadata(message),
+                    )
+
+                if message.is_final:
+                    state.final_message = message.content
+                    state.success = not message.is_error
+
+        # Check if stall was detected (CancelScope ate the Cancelled)
+        state.stalled = stall_scope.cancelled_caught
